update service manager
