+++ main/ServiceManager/ServiceManager.cpp
@@ -1,10 +1,13 @@
 #include "ServiceManager.h"
-#include "GUI.h"
-#include <stdio.h>
-#include "string.h"
 
-#include "CoffeeMakerApp.hpp"
 static const char *TAG = "Service_Manager";
+
+#ifdef CONFIG_DONE_COMPONENT_MATTER
+QueueHandle_t MatterBufQueue;
+SemaphoreHandle_t MatterSemaphore = NULL;
+MatterInterfaceHandler_t MatterInterfaceHandler;
+#endif
+
 #define TASK_LIST_BUFFER_SIZE 512
 // #define MONITORING
 /**
@@ -22,6 +25,33 @@ StackType_t *xServiceMangerStack;
 // Global instance of Service Manager
 ServiceManger_t ServiceManger;
 
+#ifdef CONFIG_DONE_COMPONENT_MATTER
+void MatterAttributeUpdateCBMain(callback_type_t type,
+                                 uint16_t endpoint_id, uint32_t cluster_id,
+                                 uint32_t attribute_id, esp_matter_attr_val_t *val,
+                                 void *priv_data)
+{
+    printf("callback_type_t: %u\n", type);
+    printf("endpoint_id: %u\n", endpoint_id);
+    printf("cluster_id: %lu\n", cluster_id);
+    printf("attribute_id: %lu\n", attribute_id);
+    printf("val: %p\n", val);
+    printf("priv_data: %pGlobalInitGlobalInitGlobalInit\n", priv_data);
+}
+
+void MatterNetworkConnected()
+{
+    ESP_LOGI(TAG, "Matter Network Connected\n");
+}
+void MatterServiceRunner()
+{
+    MatterInterfaceHandler.SharedBufQueue = &MatterBufQueue;
+    MatterInterfaceHandler.SharedSemaphore = &MatterSemaphore;
+    MatterInterfaceHandler.MatterAttributeUpdateCB = MatterAttributeUpdateCBMain;
+    MatterInterfaceHandler.ConnectToMatterNetwork = MatterNetworkConnected;
+    Matter_TaskInit(&MatterInterfaceHandler);
+}
+#endif
 #ifdef CONFIG_DONE_COMPONENT_LVGL
 /**
  * @brief Creates the GUI task.
@@ -91,13 +121,24 @@ void ServiceMangerTaskInit()
  */
 void ServiceMangerInit()
 {
+    nvsFlashInit();
+#ifndef CONFIG_DONE_COMPONENT_MATTER
+    ESP_ERROR_CHECK(esp_netif_init());
+    ESP_ERROR_CHECK(esp_event_loop_create_default());
+    ESP_ERROR_CHECK(example_connect());
+#endif
 #ifdef CONFIG_DONE_COMPONENT_LVGL
     GUI_TaskCreator();
     ESP_LOGI(TAG, "GUI Created !");
+    vTaskDelay(SERVICE_MANAGER_SEC * 4);
 #endif
 #ifdef CONFIG_DONE_COMPONENT_SPOTIFY
     SpotifyTaskCreator();
     ESP_LOGI(TAG, "Spotify Created !");
+#endif
+#ifdef CONFIG_DONE_COMPONENT_MATTER
+    MatterServiceRunner();
+    ESP_LOGI(TAG, "Matter interface called !");
 #endif
 }
 
@@ -116,9 +157,9 @@ void ServiceMangerTask(void *pvParameter)
 #endif
     while (true)
     {
-        // vTaskList(pcTaskList);
 #ifdef MONITORING
-        // ESP_LOGI(TAG, "Task List:\n%s\n", pcTaskList);
+// vTaskList(pcTaskList);
+// ESP_LOGI(TAG, "Task List:\n%s\n", pcTaskList);
 #endif
         vTaskDelay(pdMS_TO_TICKS(500));
     }
